RENAME range_sensor to Driver
both .cpp and .hpp (and Cmake reference)
